v2-2024

+++ ui/src/components/channels/ChannelsList.tsx
@@ -7,18 +7,21 @@ import { Channel } from '../../types/channels/Channel';
 import { modifyColumn } from '../../utils/modifyColumn';
 import { channelsColumnDef } from './channelsColumnDef';
 import { type MRT_ColumnDef } from 'mantine-react-table';
+import { removeEmptyKeys } from '../../utils/removeEmptyKeys';
+import { getChangedValues } from '../../utils/getChangedValues';
 
 const ChannelList = () => {
 
   const dispatch = useAppDispatch();
   const { data, loading, error } = useAppSelector((state) => state.channels)[0];
   const { currentPage } = useAppSelector((state) => state.navigation);
+  const { loggedUserId } = useAppSelector((state) => state.session);
+
   const initialState = {
     showColumnFilters: false,
     showGlobalFilter: true,
     columnVisibility: {
       id: false,
-      password: false,
     },
   }
 
@@ -26,24 +29,37 @@ const ChannelList = () => {
     dispatch(fetchChannels())
   }, [dispatch]);
 
-  const handleCreate = async (dataCreate: Channel) => {
-    await dispatch(createChannel(dataCreate));
-    dispatch(fetchChannels());
+  const handleCreate = async (dataCreate: Partial<Channel>) => {
+    const dataCreated = removeEmptyKeys(dataCreate, loggedUserId);
+    if(Object.keys(dataCreated).some(key => key !== 'createdBy') && Object.keys(dataCreated).length !== 0){
+      await dispatch(createChannel(dataCreated));
+      dispatch(fetchChannels());
+    }
   };
 
-  const handleUpdate = async (dataUpdate: Channel) => {
-    const channelId = dataUpdate.id;
-    const channelData = dataUpdate;
-    await dispatch(updateChannel({ channelId, channelData }));
-    dispatch(fetchChannels());
+  const handleUpdate = async (originalData: Partial<Channel>, dataUpdated: Partial<Channel>) => {
+    const dataId = originalData.id;
+    const dataUpdate = getChangedValues(originalData, dataUpdated, loggedUserId);
+    if (typeof dataId === 'number') {
+      if(Object.keys(dataUpdate).some(key => key !== 'updatedBy') && Object.keys(dataUpdate).length !== 0){
+        await dispatch(updateChannel({ channelId:dataId, channelData:dataUpdate }));
+        dispatch(fetchChannels());
+      }
+    }else{
+      console.error('Channel ID is undefined.');
+    }
   };
 
-  const handleDelete = async (dataDelete: Channel, count: number, iterator: number) => {
-    await dispatch(deleteChannel(dataDelete.id));
-    if (count === iterator) { dispatch(fetchChannels()); }
+  const handleDelete = async (dataDelete: Partial<Channel>, count: number, iterator: number) => {
+    if (typeof dataDelete.id === 'number') {
+      await dispatch(deleteChannel(dataDelete.id));
+      if (count === iterator) { dispatch(fetchChannels()); }
+    }else{
+      console.error('Channel ID is undefined.');
+    }
   };
 
-  const modifiedColumns = useMemo<MRT_ColumnDef<Channel>[]>(() => modifyColumn(data[0]?.columns || [], channelsColumnDef), [data]);
+  const modifiedColumns = useMemo<MRT_ColumnDef<Partial<Channel>>[]>(() => modifyColumn(data[0]?.columns || [], channelsColumnDef), [data]);
 
   return (
     <Table
